third_party/WebKit/LayoutTests/inspector/console/console-format-table-expected.txt

CONSOLE MESSAGE: line 8: null
CONSOLE MESSAGE: line 10: Array of arrays
CONSOLE MESSAGE: line 11: 1,2,3,4,5,6
CONSOLE MESSAGE: line 13: Large array of arrays
CONSOLE MESSAGE: line 14: 1,2,3,4,5,6,1,2,3,4,5,6,1,2,3,4,5,6,1,2,3,4,5,6
CONSOLE MESSAGE: line 16: Array or array and object
CONSOLE MESSAGE: line 17: 1,2,3,[object Object]
CONSOLE MESSAGE: line 19: Object table
CONSOLE MESSAGE: line 20: [object Object]
CONSOLE MESSAGE: line 22: Null as columns
CONSOLE MESSAGE: line 23: 1,2,3,4,5,6
CONSOLE MESSAGE: line 25: Digit as columns
CONSOLE MESSAGE: line 26: 1,2,3,4,5,6
CONSOLE MESSAGE: line 28: String as columns
CONSOLE MESSAGE: line 29: 1,2,3,4,5,6
CONSOLE MESSAGE: line 31: Random string as columns
CONSOLE MESSAGE: line 32: 1,2,3,4,5,6
CONSOLE MESSAGE: line 34: Array of strings as columns
CONSOLE MESSAGE: line 35: [object Object],[object Object]
CONSOLE MESSAGE: line 37: Good and bad column names
CONSOLE MESSAGE: line 38: [object Object],[object Object]
CONSOLE MESSAGE: line 40: Missing column name
CONSOLE MESSAGE: line 41: [object Object],[object Object]
CONSOLE MESSAGE: line 43: Shallow array
CONSOLE MESSAGE: line 44: 1,foo,
CONSOLE MESSAGE: line 46: Shallow array with 'Value' column
CONSOLE MESSAGE: line 47: 1,[object Object]
CONSOLE MESSAGE: line 49: Deep and shallow array
CONSOLE MESSAGE: line 50: 1,foo,2
CONSOLE MESSAGE: line 8: null
CONSOLE MESSAGE: line 10: Array of arrays
CONSOLE MESSAGE: line 11: 1,2,3,4,5,6
CONSOLE MESSAGE: line 13: Large array of arrays
CONSOLE MESSAGE: line 14: 1,2,3,4,5,6,1,2,3,4,5,6,1,2,3,4,5,6,1,2,3,4,5,6
CONSOLE MESSAGE: line 16: Array or array and object
CONSOLE MESSAGE: line 17: 1,2,3,[object Object]
CONSOLE MESSAGE: line 19: Object table
CONSOLE MESSAGE: line 20: [object Object]
CONSOLE MESSAGE: line 22: Null as columns
CONSOLE MESSAGE: line 23: 1,2,3,4,5,6
CONSOLE MESSAGE: line 25: Digit as columns
CONSOLE MESSAGE: line 26: 1,2,3,4,5,6
CONSOLE MESSAGE: line 28: String as columns
CONSOLE MESSAGE: line 29: 1,2,3,4,5,6
CONSOLE MESSAGE: line 31: Random string as columns
CONSOLE MESSAGE: line 32: 1,2,3,4,5,6
CONSOLE MESSAGE: line 34: Array of strings as columns
CONSOLE MESSAGE: line 35: [object Object],[object Object]
CONSOLE MESSAGE: line 37: Good and bad column names
CONSOLE MESSAGE: line 38: [object Object],[object Object]
CONSOLE MESSAGE: line 40: Missing column name
CONSOLE MESSAGE: line 41: [object Object],[object Object]
CONSOLE MESSAGE: line 43: Shallow array
CONSOLE MESSAGE: line 44: 1,foo,
CONSOLE MESSAGE: line 46: Shallow array with 'Value' column
CONSOLE MESSAGE: line 47: 1,[object Object]
CONSOLE MESSAGE: line 49: Deep and shallow array
CONSOLE MESSAGE: line 50: 1,foo,2
Tests console.table.

console-format-table.html:8 
console-format-table.html:10 Array of arrays
console-format-table.html:11 
console-format-table.html:13 Large array of arrays
console-format-table.html:14 
console-format-table.html:16 Array or array and object
console-format-table.html:17 
console-format-table.html:19 Object table
console-format-table.html:20 
console-format-table.html:22 Null as columns
console-format-table.html:23 
console-format-table.html:25 Digit as columns
console-format-table.html:26 
console-format-table.html:28 String as columns
console-format-table.html:29 
console-format-table.html:31 Random string as columns
console-format-table.html:32 
console-format-table.html:34 Array of strings as columns
console-format-table.html:35 
console-format-table.html:37 Good and bad column names
console-format-table.html:38 
console-format-table.html:40 Missing column name
console-format-table.html:41 
console-format-table.html:43 Shallow array
console-format-table.html:44 
console-format-table.html:46 Shallow array with 'Value' column
console-format-table.html:47 
console-format-table.html:49 Deep and shallow array
console-format-table.html:50 
console-format-table.html:8 
console-format-table.html:10 Array of arrays
HEADER (index) | 0 | 1 | 2 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
console-format-table.html:13 Large array of arrays
HEADER (index) | 0 | 1 | 2 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
console-format-table.html:16 Array or array and object
HEADER (index) | 0 | 1 | 2 | a | b | c | 
ROW 1 | 2 | 3 | 
ROW 1 | 2 | 3 | 
console-format-table.html:19 Object table
HEADER (index) | a | b | 
ROW 1 | 2 | 
ROW 3 | 4 | 
console-format-table.html:22 Null as columns
HEADER (index) | 0 | 1 | 2 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
console-format-table.html:25 Digit as columns
HEADER (index) | 0 | 1 | 2 | 
ROW 1 | 2 | 3 | 
ROW 4 | 5 | 6 | 
console-format-table.html:28 String as columns
HEADER (index) | 0 | 
ROW 1 | 
ROW 4 | 
console-format-table.html:31 Random string as columns
HEADER (index) | 
console-format-table.html:34 Array of strings as columns
HEADER (index) | a | b | 
ROW 1 | 2 | 
ROW "foo" | "bar" | 
console-format-table.html:37 Good and bad column names
HEADER (index) | a | b | 
ROW 1 | 2 | 
ROW "foo" | 
console-format-table.html:40 Missing column name
HEADER (index) | 
console-format-table.html:43 Shallow array
HEADER (index) | Value | 
ROW 1 | 
ROW "foo" | 
ROW null | 
console-format-table.html:46 Shallow array with 'Value' column
HEADER (index) | Value | Value | 
ROW 1 | 
ROW 2 | 
console-format-table.html:49 Deep and shallow array
HEADER (index) | Value | 0 | 
ROW 1 | 
ROW "foo" | 
ROW 2 | 
